Added modals and a new settings modal
The new Modal component provides a "modal"-style element that coverts the whole page.
Using `inert`, `aria-hidden`, and just covering the page, a modal also steals focus from the other page elements.
The ModalManager keeps track of all open modals and detects what to give focus too (even if only one modal should be displayed at a time).
Finally, as a test, the SettingsModal has been added and a button to trigger it has been placed in the header.